update: rename calendar component file
